Move ApiClient and DeferredJobsApiClient to a separate folder
