+++ src/Sdk/Accounts/Accounts.ts
@@ -1,5 +1,5 @@
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 import { UserAccount } from '../../types';
 import {
     AccountsCreateRequest,
