+++ src/Sdk/Client.ts
@@ -1,10 +1,9 @@
+import { ApiClient, DeferredJobsApiClient } from '../api';
 import { ClientOptions } from './types';
-import { ApiClient } from './ApiClient';
 import { Accounts } from './Accounts';
 import Campaigns from './Campaigns';
 import CampaignRecipients from './CampaignRecipients';
 import Coverage from './Coverage';
-import { DeferredJobsApiClient } from './DeferredJobsApiClient';
 import Stories from './Stories';
 import Snippets from './Snippets';
 import Newsrooms from './Newsrooms';
