+++ src/Sdk/Jobs/Jobs.ts
@@ -1,7 +1,7 @@
 import { JobState } from '../../types';
 
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 
 import { JobStatusResponse } from './types';
