+++ src/Sdk/NewsroomCategories/NewsroomCategories.ts
@@ -1,7 +1,7 @@
 import { Category, Newsroom } from '../../types';
 
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 
 import {
     NewsroomCategoriesListOptions,
