+++ src/Sdk/NewsroomContacts/NewsroomContacts.ts
@@ -1,7 +1,7 @@
 import { Newsroom, NewsroomContact } from '../../types';
 
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 
 import {
     NewsroomContactsListRequestOptions,
