+++ src/Sdk/NewsroomGalleries/NewsroomGalleries.ts
@@ -1,7 +1,7 @@
 import { Newsroom } from '../../types';
 
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 import { NewsroomGallery } from '../../types';
 import {
     NewsroomGalleriesListRequest,
