+++ src/Sdk/NewsroomPrivacyRequests/NewsroomPrivacyRequests.ts
@@ -1,7 +1,7 @@
 import { Newsroom, PrivacyRequest } from '../../types';
 
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
-import routing from '../routing';
+import { DeferredJobsApiClient } from '../../api';
+import { routing } from '../../routing';
 import {
     DeletePrivacyRequestCreateRequest,
     ExportPrivacyRequestCreateRequest,
