+++ src/Sdk/NewsroomSubscriptions/NewsroomSubscriptions.ts
@@ -1,7 +1,7 @@
 import { Newsroom } from '../../types';
 
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
-import routing from '../routing';
+import { DeferredJobsApiClient } from '../../api';
+import { routing } from '../../routing';
 
 import { NewsroomSubscriptionCreateRequest } from './types';
