+++ src/Sdk/NewsroomThemes/NewsroomThemes.ts
@@ -1,7 +1,7 @@
 import { Newsroom, NewsroomTheme, NewsroomThemePreset } from '../../types';
 
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 
 type NewsroomId = Newsroom['uuid'] | Newsroom['id'];
 type NewsroomThemeId = NewsroomTheme['id'];
