+++ src/Sdk/Stories/Stories.ts
@@ -1,7 +1,7 @@
 import { ExtendedStory, ExtraStoryFields, Story } from '../../types';
 
-import routing from '../routing';
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
+import { routing } from '../../routing';
+import { DeferredJobsApiClient } from '../../api';
 
 import {
     StoriesListRequest,
