+++ src/Sdk/Subscriptions/Subscriptions.ts
@@ -1,7 +1,7 @@
 import { ContactType, EmailSubscription, Newsroom } from '../../types';
 
-import { DeferredJobsApiClient } from '../DeferredJobsApiClient';
-import routing from '../routing';
+import { DeferredJobsApiClient } from '../../api';
+import { routing } from '../../routing';
 
 import { SubscribeRequest, UnsubscribeRequest, UpdateUnsubscribeDetailsRequest } from './types';
